Move triangular solvers to combined header
