+++ include/ginkgo/ginkgo.hpp
@@ -125,6 +125,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <ginkgo/core/solver/multigrid.hpp>
 #include <ginkgo/core/solver/solver_base.hpp>
 #include <ginkgo/core/solver/solver_traits.hpp>
+#include <ginkgo/core/solver/triangular.hpp>
 #include <ginkgo/core/solver/upper_trs.hpp>
 #include <ginkgo/core/solver/workspace.hpp>
